+++ tasks.py
@@ -1,12 +1,11 @@
 from invoke import Collection
 
-from invocations import docs
-from invocations.checks import blacken
+from invocations import docs, checks
 from invocations.packaging import release
 from invocations.pytest import test, coverage
 
 
-ns = Collection(release, test, coverage, docs, blacken)
+ns = Collection(release, test, coverage, docs, checks.blacken, checks)
 ns.configure(
     {
         "packaging": {